fix: disable irrelevant informations (#2910)
The "PlayerBadge" was loading a lot of unnecessary information for all
players on the account. I moved the if statement up, ensuring that only
the preload and level are properly loaded, which are actually relevant
information.

If more information is identified that should be loaded in the preload,
we should move it to the loadPlayerFirst function.